Sentiment_Analysis/requirements.txt
streamlit
google-generativeai
python-dotenv
streamlit-lottie
requests